LCDロジック分離
