+++ RetsubanWindow/RetsubanResource.resx
@@ -337,13 +337,13 @@
   <data name="_16dot_CY" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>Image\16dot_CY.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
-  <data name="LCD_Font" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>Image\LCD_Font.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
   <data name="_7seg_1" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>Image\7seg_1.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
   <data name="Button_D" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>Image\Button_D.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="LCD_Font" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>Image\LCD_Font.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
 </root>
